BAU: Refactor - move CRI response code into criresponse namespace
