+++ libs/cri-response-service/src/main/java/uk/gov/di/ipv/core/library/criresponse/service/CriResponseService.java
@@ -1,14 +1,15 @@
-package uk.gov.di.ipv.core.library.service;
+package uk.gov.di.ipv.core.library.criresponse.service;
 
 import com.nimbusds.jwt.util.DateUtils;
 import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
 import uk.gov.di.ipv.core.library.config.ConfigurationVariable;
-import uk.gov.di.ipv.core.library.domain.AsyncCriStatus;
+import uk.gov.di.ipv.core.library.criresponse.domain.AsyncCriStatus;
 import uk.gov.di.ipv.core.library.domain.Cri;
 import uk.gov.di.ipv.core.library.domain.VerifiableCredential;
 import uk.gov.di.ipv.core.library.persistence.DataStore;
 import uk.gov.di.ipv.core.library.persistence.item.ClientOAuthSessionItem;
 import uk.gov.di.ipv.core.library.persistence.item.CriResponseItem;
+import uk.gov.di.ipv.core.library.service.ConfigService;
 
 import java.time.Instant;
 import java.util.Date;
